+++ postgres-derive/src/tosql.rs
@@ -1,62 +1,95 @@
-use proc_macro2::TokenStream;
+use proc_macro2::{Span, TokenStream};
 use quote::quote;
 use std::iter;
-use syn::{Data, DataStruct, DeriveInput, Error, Fields, Ident};
+use syn::{
+    Data, DataStruct, DeriveInput, Error, Fields, Ident, TraitBound, TraitBoundModifier,
+    TypeParamBound,
+};
 
 use crate::accepts;
 use crate::composites::Field;
+use crate::composites::{append_generic_bound, new_derive_path};
 use crate::enums::Variant;
 use crate::overrides::Overrides;
 
 pub fn expand_derive_tosql(input: DeriveInput) -> Result<TokenStream, Error> {
     let overrides = Overrides::extract(&input.attrs)?;
 
+    if overrides.name.is_some() && overrides.transparent {
+        return Err(Error::new_spanned(
+            &input,
+            "#[postgres(transparent)] is not allowed with #[postgres(name = \"...\")]",
+        ));
+    }
+
     let name = overrides.name.unwrap_or_else(|| input.ident.to_string());
 
-    let (accepts_body, to_sql_body) = match input.data {
-        Data::Enum(ref data) => {
-            let variants = data
-                .variants
-                .iter()
-                .map(Variant::parse)
-                .collect::<Result<Vec<_>, _>>()?;
-            (
-                accepts::enum_body(&name, &variants),
-                enum_body(&input.ident, &variants),
-            )
+    let (accepts_body, to_sql_body) = if overrides.transparent {
+        match input.data {
+            Data::Struct(DataStruct {
+                fields: Fields::Unnamed(ref fields),
+                ..
+            }) if fields.unnamed.len() == 1 => {
+                let field = fields.unnamed.first().unwrap();
+
+                (accepts::transparent_body(field), transparent_body())
+            }
+            _ => {
+                return Err(Error::new_spanned(
+                    input,
+                    "#[postgres(transparent)] may only be applied to single field tuple structs",
+                ));
+            }
         }
-        Data::Struct(DataStruct {
-            fields: Fields::Unnamed(ref fields),
-            ..
-        }) if fields.unnamed.len() == 1 => {
-            let field = fields.unnamed.first().unwrap();
-            (accepts::domain_body(&name, field), domain_body())
-        }
-        Data::Struct(DataStruct {
-            fields: Fields::Named(ref fields),
-            ..
-        }) => {
-            let fields = fields
-                .named
-                .iter()
-                .map(Field::parse)
-                .collect::<Result<Vec<_>, _>>()?;
-            (
-                accepts::composite_body(&name, "ToSql", &fields),
-                composite_body(&fields),
-            )
-        }
-        _ => {
-            return Err(Error::new_spanned(
-                input,
-                "#[derive(ToSql)] may only be applied to structs, single field tuple structs, and enums",
-            ));
+    } else {
+        match input.data {
+            Data::Enum(ref data) => {
+                let variants = data
+                    .variants
+                    .iter()
+                    .map(Variant::parse)
+                    .collect::<Result<Vec<_>, _>>()?;
+                (
+                    accepts::enum_body(&name, &variants),
+                    enum_body(&input.ident, &variants),
+                )
+            }
+            Data::Struct(DataStruct {
+                fields: Fields::Unnamed(ref fields),
+                ..
+            }) if fields.unnamed.len() == 1 => {
+                let field = fields.unnamed.first().unwrap();
+
+                (accepts::domain_body(&name, field), domain_body())
+            }
+            Data::Struct(DataStruct {
+                fields: Fields::Named(ref fields),
+                ..
+            }) => {
+                let fields = fields
+                    .named
+                    .iter()
+                    .map(Field::parse)
+                    .collect::<Result<Vec<_>, _>>()?;
+                (
+                    accepts::composite_body(&name, "ToSql", &fields),
+                    composite_body(&fields),
+                )
+            }
+            _ => {
+                return Err(Error::new_spanned(
+                    input,
+                    "#[derive(ToSql)] may only be applied to structs, single field tuple structs, and enums",
+                ));
+            }
         }
     };
 
     let ident = &input.ident;
+    let generics = append_generic_bound(input.generics.to_owned(), &new_tosql_bound());
+    let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
     let out = quote! {
-        impl postgres_types::ToSql for #ident {
+        impl#impl_generics postgres_types::ToSql for #ident#ty_generics #where_clause {
             fn to_sql(&self,
                       _type: &postgres_types::Type,
                       buf: &mut postgres_types::private::BytesMut)
@@ -76,6 +109,12 @@ pub fn expand_derive_tosql(input: DeriveInput) -> Result<TokenStream, Error> {
     };
 
     Ok(out)
+}
+
+fn transparent_body() -> TokenStream {
+    quote! {
+        postgres_types::ToSql::to_sql(&self.0, _type, buf)
+    }
 }
 
 fn enum_body(ident: &Ident, variants: &[Variant]) -> TokenStream {
@@ -147,4 +186,13 @@ fn composite_body(fields: &[Field]) -> TokenStream {
 
         std::result::Result::Ok(postgres_types::IsNull::No)
     }
+}
+
+fn new_tosql_bound() -> TypeParamBound {
+    TypeParamBound::Trait(TraitBound {
+        lifetimes: None,
+        modifier: TraitBoundModifier::None,
+        paren_token: None,
+        path: new_derive_path(Ident::new("ToSql", Span::call_site()).into()),
+    })
 }